+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 #### TBD
 - Update to new base URL, fix #24
 - Repair and rewrite group crawling
+- Workaround robobrowser bug for Python 3.7, fix #25
 
 ## [0.63] (https://github.com/voyageur/dagr/tree/0.63)
 #### 02/01/2018